Re-add config, fallback to example config
I made it so that if a config.json file isn't found, we fall
back to the example.config.json file. This will keep development
possible without needing API keys, so long as we implement API
calls correctly (verify key exists before trying). It will output
to the console for easy debugging when no config at all is found.
The failure cases for the http.gets are now being handled